Adopt UIScene lifecycle in UIKit example app
