+++ Examples/UIKit/SBTUITestTunnel/SceneDelegate.swift
@@ -1,0 +1,31 @@
+// SceneDelegate.swift
+//
+// Copyright (C) 2016 Subito.it S.r.l (www.subito.it)
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+import UIKit
+
+class SceneDelegate: UIResponder, UIWindowSceneDelegate {
+    var window: UIWindow?
+
+    func scene(_ scene: UIScene, willConnectTo _: UISceneSession, options _: UIScene.ConnectionOptions) {
+        guard let windowScene = scene as? UIWindowScene else { return }
+
+        window = UIWindow(windowScene: windowScene)
+        let rootVC = SBTTableViewController(style: .plain)
+        let navController = UINavigationController(rootViewController: rootVC)
+        window?.rootViewController = navController
+        window?.makeKeyAndVisible()
+    }
+}
